Convert AuctionBoard to HiringBoard
