--- src/views/AuctionQueue/components/index.ts
@@ -1,2 +1,0 @@
-export { HiringBoard } from './HiringBoard';
-export { OpenBounty } from './OpenBounty';
