+++ src/views/HiringBoard/AuctionQueue.scss
